+++ MassiveSearchBundle.php
@@ -1,6 +1,7 @@
 <?php
+
 /*
- * This file is part of the Sulu CMS.
+ * This file is part of the MassiveSearchBundle
  *
  * (c) MASSIVE ART WebServices GmbH
  *
@@ -13,6 +14,7 @@ namespace Massive\Bundle\SearchBundle;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 use Massive\Bundle\SearchBundle\DependencyInjection\Compiler\MetadataDriverPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Massive\Bundle\SearchBundle\DependencyInjection\Compiler\MetadataProviderPass;
 
 class MassiveSearchBundle extends Bundle
 {
@@ -20,5 +22,6 @@ class MassiveSearchBundle extends Bundle
     {
         parent::build($container);
         $container->addCompilerPass(new MetadataDriverPass());
+        $container->addCompilerPass(new MetadataProviderPass());
     }
 }