+++ Search/SearchManager.php
@@ -1,6 +1,7 @@
 <?php
+
 /*
- * This file is part of the Sulu CMS.
+ * This file is part of the MassiveSearchBundle
  *
  * (c) MASSIVE ART WebServices GmbH
  *
@@ -10,13 +11,14 @@
 
 namespace Massive\Bundle\SearchBundle\Search;
 
-use Massive\Bundle\SearchBundle\Search\Event\SearchEvent;
-use Massive\Bundle\SearchBundle\Search\Metadata\IndexMetadata;
 use Massive\Bundle\SearchBundle\Search\Event\HitEvent;
-use Metadata\MetadataFactory;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Massive\Bundle\SearchBundle\Search\Event\PreIndexEvent;
+use Massive\Bundle\SearchBundle\Search\Event\SearchEvent;
 use Massive\Bundle\SearchBundle\Search\Exception\MetadataNotFoundException;
+
+use Massive\Bundle\SearchBundle\Search\Metadata\ProviderInterface as MetadataProviderInterface;
+
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
  * Search manager is the public API to the search
@@ -30,12 +32,12 @@ class SearchManager implements SearchManagerInterface
     protected $adapter;
 
     /**
-     * @var \Metadata\MetadataFactory
+     * @var ProviderInterface
      */
-    protected $metadataFactory;
+    protected $metadataProvider;
 
     /**
-     * @var \Symfony\Component\EventDispatcher\EventDispatcherInterface
+     * @var EventDispatcherInterface
      */
     protected $eventDispatcher;
 
@@ -56,13 +58,13 @@ class SearchManager implements SearchManagerInterface
 
     public function __construct(
         AdapterInterface $adapter,
-        MetadataFactory $metadataFactory,
+        MetadataProviderInterface $metadataProvider,
         ObjectToDocumentConverter $converter,
         EventDispatcherInterface $eventDispatcher,
         LocalizationStrategyInterface $localizationStrategy
     ) {
         $this->adapter = $adapter;
-        $this->metadataFactory = $metadataFactory;
+        $this->metadataProvider = $metadataProvider;
         $this->eventDispatcher = $eventDispatcher;
         $this->converter = $converter;
         $this->localizationStrategy = $localizationStrategy;
@@ -83,7 +85,18 @@ class SearchManager implements SearchManagerInterface
             );
         }
 
-        return $this->getMetadataFor(get_class($object));
+        $metadata = $this->metadataProvider->getMetadataForObject($object);
+
+        if (null === $metadata) {
+            throw new MetadataNotFoundException(
+                sprintf(
+                    'There is no search mapping for object with class "%s"',
+                    get_class($object)
+                )
+            );
+        }
+
+        return $metadata;
     }
 
     /**
@@ -109,9 +122,9 @@ class SearchManager implements SearchManagerInterface
      */
     public function index($object)
     {
-        $indexMetadatas = $this->getMetadata($object);
+        $indexMetadata = $this->getMetadata($object);
 
-        foreach ($indexMetadatas->getIndexMetadatas() as $indexMetadata) {
+        foreach ($indexMetadata->getIndexMetadatas() as $indexMetadata) {
             $this->markIndexToFlush($indexMetadata->getIndexName());
 
             $indexName = $indexMetadata->getIndexName();
@@ -149,6 +162,15 @@ class SearchManager implements SearchManagerInterface
         $this->validateQuery($query);
         $this->expandQueryIndexes($query);
 
+        // At this point the indexes should have been expanded to potentially
+        // include all indexes managed by massive search, if it is empty then
+        // there is nothing to search for.
+        //
+        // See: https://github.com/massiveart/MassiveSearchBundle/issues/38
+        if (0 === count($query->getIndexes())) {
+            return array();
+        }
+
         $this->eventDispatcher->dispatch(
             SearchEvents::SEARCH,
             new SearchEvent($query)
@@ -165,7 +187,7 @@ class SearchManager implements SearchManagerInterface
                 continue;
             }
 
-            $metadata = $this->getMetadataFor($document->getClass());
+            $metadata = $this->metadataProvider->getMetadataForDocument($document);
             $indexMetadata = $metadata->getIndexMetadata('_default');
             $document->setCategory($indexMetadata->getCategoryName());
 
@@ -184,7 +206,7 @@ class SearchManager implements SearchManagerInterface
     public function getStatus()
     {
         $data = array('Adapter' => get_class($this->adapter));
-        $data += $this->adapter->getStatus() ? : array();
+        $data += $this->adapter->getStatus() ?: array();
 
         return $data;
     }
@@ -206,12 +228,10 @@ class SearchManager implements SearchManagerInterface
      */
     public function getCategoryNames()
     {
-        $classNames = $this->metadataFactory->getAllClassNames();
+        $metadatas = $this->metadataProvider->getAllMetadata();
         $categoryNames = array();
 
-        foreach ($classNames as $className) {
-            $metadata = $this->getMetadataFor($className);
-
+        foreach ($metadatas as $metadata) {
             foreach ($metadata->getIndexMetadatas() as $indexMetadata) {
                 $categoryNames[] = $indexMetadata->getCategoryName();
             }
@@ -230,7 +250,7 @@ class SearchManager implements SearchManagerInterface
     }
 
     /**
-     * Return a list of all the index names (according to the metadata)
+     * Return a list of all the index names (according to the metadata).
      *
      * If categories are specified, only return the indexes corresponding
      * to the given categories.
@@ -241,12 +261,10 @@ class SearchManager implements SearchManagerInterface
      */
     public function getIndexNames($categories = null)
     {
-        $classNames = $this->metadataFactory->getAllClassNames();
+        $metadatas = $this->metadataProvider->getAllMetadata();
         $indexNames = array();
 
-        foreach ($classNames as $className) {
-            $metadata = $this->getMetadataFor($className);
-
+        foreach ($metadatas as $metadata) {
             foreach ($metadata->getIndexMetadatas() as $indexMetadata) {
                 $indexName = $indexMetadata->getIndexName();
                 if ($categories) {
@@ -268,6 +286,7 @@ class SearchManager implements SearchManagerInterface
      * optionally only in the given locale.
      *
      * @param string $locale
+     *
      * @return string[]
      */
     private function getLocalizedIndexNames($locale = null)
@@ -330,6 +349,7 @@ class SearchManager implements SearchManagerInterface
         if (!$query->getIndexes()) {
             $indexNames = $this->getLocalizedIndexNames($query->getLocale());
             $query->setIndexes($indexNames);
+
             return;
         }
 
@@ -353,32 +373,10 @@ class SearchManager implements SearchManagerInterface
     }
 
     /**
-     * Return metadata for the given classname
-     *
-     * @param string $className
-     *
-     * @return ClassMetadata
-     */
-    private function getMetadataFor($className)
-    {
-        $metadata = $this->metadataFactory->getMetadataForClass($className);
-
-        if (null === $metadata) {
-            throw new MetadataNotFoundException(
-                sprintf(
-                    'There is no search mapping for class "%s"',
-                    $className
-                )
-            );
-        }
-
-        return $metadata->getOutsideClassMetadata();
-    }
-
-    /**
-     * If query has indexes, ensure that they are known
+     * If query has indexes, ensure that they are known.
      *
      * @throws Exception\SearchException
+     *
      * @param SearchQuery $query
      */
     private function validateQuery(SearchQuery $query)