+++ Tests/Resources/TestBundle/EntityDist/Product.php
@@ -1,6 +1,16 @@
 <?php
 
+/*
+ * This file is part of the MassiveSearchBundle
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Massive\Bundle\SearchBundle\Tests\Resources\TestBundle\Entity;
+
 use Massive\Bundle\SearchBundle\Tests\Resources\TestBundle\Product as BaseProduct;
 
 class Product extends BaseProduct