+++ Tests/Unit/DependencyInjection/MassiveSearchExtensionTest.php
@@ -1,6 +1,7 @@
 <?php
+
 /*
- * This file is part of the Sulu CMS.
+ * This file is part of the MassiveSearchBundle
  *
  * (c) MASSIVE ART WebServices GmbH
  *
@@ -20,6 +21,8 @@ class MassiveSearchExtensionTest extends AbstractExtensionTestCase
         parent::setUp();
 
         $this->container->setParameter('kernel.root_dir', '/some/path');
+        $this->container->setParameter('kernel.cache_dir', __DIR__ . '/../../Resources/app/cache');
+        $this->container->setParameter('kernel.debug', false);
         $this->container->register('event_dispatcher', 'Symfony\Component\EventDispatcher\EventDispatcher');
     }